3.4.0 Instead of sending the welcome message when the user joins send it when the Member role is added
sorry for anyone who wants to copy this code but doesn't need that capability
