+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "jaiden-animations-discord-utilities",
-  "version": "3.3.0",
+  "version": "3.4.0",
   "description": "Discord bot with utilities for r/JaidenAnimations Discord Server",
   "main": "index.js",
   "scripts": {},
